Switch postgresql.client away from pgbouncer
PGBouncer does not support connecting to the PostgreSQL server using
TLS. They recommend using stunnel to handle this instead however
stunnel does not support verifying the common name of the certificate.
This means that in order to get proper full verification of the
authenticity of the PostgreSQL server we cannot use pgbouncer.

In addition the use of pgbouncer is likely a premature optimization
so simply dropping it in favor of direct connections makes the
most sense.