Disable the TLS Session Tickets (Thanks @meineerde)
The way TLS Session Tickets are handled in HAProxy is a key is
generated when the process starts up and is never recycled until
the process restarts again. This means that if an attacker gains
access to this TLS Session Ticket they can decrypt all of the
communications that were protected with that key. TLS Session IDs
on the other hand have a finite lifetime (~5 minutes currently).

Disabling this will ensure that our PFS stays PFS.