Use the PSF_CA and only give certificates to the consul servers
Consul (really go's TLS) does not check the extendKeyUsage x509
extension to ensure that a certificate is valid for serverAuth.
This means that if a client was compromised they could turn themselves
into a server. Until Consul supports either a different server_name
for server and client, or it supports checking for a serverAuth in
the extendedKeyUsage we need to restrict TLS to only the servers
ensuring that a client won't have a TLS certificate that is acceptable
to be a server.

We also configure server_name so that we can reuse the PSF_CA with
a certificate with the CN of consul.psf.io instead of relying on
a seperate CA.