Restrict to a subnet, not to specific IP addresses
Originally this was designed so that a leaked password is only
useful on the system it was designed for. However doing this relies
on mine.get in order for it to actually provide any sort of security
and mine.get takes a long time to update.

This will still prevent people from *other* networks being able
to access the databases and we can investigate using TLS client
certificates instead of passwords as well.