Use stunnel instead of PostgreSQL to terminate TLS
PostgreSQL doesn't offer the ability to gracefully reload the TLS
certificate, which requires a hard restart whenever the certificate
is close to expiring. This generates errors because it interupts
current connections. Instead, we'll utilize stunnel to terminate
TLS which does support a graceful reload.

The only major downsides to this, is PostgreSQL isn't aware that
the connections are happening via TLS and since we're proxying
connections, PostgreSQL sees all connections as coming from
127.0.0.1 instead of the real source address.