pebble-source state failing (#301)
When bringing up the salt-master locally, the pebble-source state `git.latest` function fails with this comment `Repository would be updated from 087582e to d5fa738, but there are uncommitted changes. Set 'force_reset' to True (or 'remote-changes') to force this update and discard these changes.`
By defualt, git.lastest force_reset is set to false. By changing our configuration to use `force_reset: remote-changes`, we can instruct salt not to discard local changes if the repo is up-to-date with the remote repository.