configure mod_qos
this configures mod_qos globally to use the client level control directives `QS_ClientEvent` to implement a harsh rate limit for testing purposes and produce a 429 error when the limit exceeds. it also sets an environmental variable `QS_Event` that points to our `Request_URI` triggering the `rate_limit` when the limit count exceeds.