Restructure our Dockerfiles and incorporate 22.04 as an option
This restructures us from a single Dockerfile to multiple allowing
us to have more than one active distribution at a time.

In order to "upgrade" a given vagrant box, configure the `codename`
value in the Vagrant file for that box to the matching distro codename.

Salt-master is special cased as it is configured in-line.
Uncomment the current bit or update the "build_dir" for its docker provider
in order to change its distro.