Run salt-master as root user
In our previoius configuration, the salt master was running as root user. When migrating to salt latest 3006, the salt-master defaults to running as the salt user.
To bypass permissions errors around loading `consul.py` and `ca.py`, this commit runs the salt-master as the root user as done previously before attempting to upgrade salt.