Migrate salt-master to Ubuntu 22.04('Jammy') (#331)
* Migrate salt-master to jammy

This commit uncomments out code that will migrate our salt-master to run using Ubuntu 22.04 codename 'jammy'.

* Run salt-master as root user

In our previoius configuration, the salt master was running as root user. When migrating to salt latest 3006, the salt-master defaults to running as the salt user.
To bypass permissions errors around loading `consul.py` and `ca.py`, this commit runs the salt-master as the root user as done previously before attempting to upgrade salt.

* Expand condtional statement to include codename "jammy"

The `python-requests` and `python-msgpack` packages were not getting loaded for our salt-master because there was no condtion to match it's oscodename. This commit changes that.

* Upadte saltstack repository configuration for codename "jammy"

This commit applies a condition for codename jammy to use the appropriate salt repository, updating the repo URL, path name, and GPG key for Ubuntu 22.04.

* Adjust `consul` to run with jammy

Adjust `consul` configuration for compatibility with Ubuntu 22.04 ('Jammy')

* Remove comment

This comment is no longer necessary, as the salt-master now runs on Ubuntu 22.04 codename "Jammy".

* Update salt GPG keyring name

- this change reflects the GPG keyring name, and apt source file list in [Salt Install Guide]https://docs.saltproject.io/salt/install-guide/en/latest/topics/install-by-operating-system/ubuntu.html)

* Address apt-key deprecation for packagecloud

- this commit addresses the recommended approach for [apt-key deprecation in Ubunut 22.04](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.pkgrepo.html)

* Address apt-key deprecation for salt

- this commit addresses the reccomended approach for  [apt-key deprecation in Ubunut 22.04](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.pkgrepo.html)

* Address apt-key deprecation for datadoghq

- this commit addresses the reccomended approach for  [apt-key deprecation in Ubunut 22.04](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.pkgrepo.html)

* Address apt-key deprecation for nginx

- this commit addresses the reccomended approach for  [apt-key deprecation in Ubunut 22.04](https://docs.saltproject.io/en/latest/ref/states/all/salt.states.pkgrepo.html)