Allow Consul DNS to recurse on haproxy nodes
PR #486 introduced resolution of hosts for haproxy backends via consul, but did not give it any way to resolve external DNS for the console/pythonanywhere or jobs.pyfound.org/applytojob.com backends.

This adds a recursive DNS setup on the haproxy nodes so that they can successfully and fully resolve external services.